requirements.txt
pandas
premailer
pyyaml
stevedore
# [mysql]
mysqlclient